rename assertion to table/utils (as it's only used there)
