+++ static_report/src/components/shared/Tables/TableList/CRTableListItem/CRTableListColumnList.tsx
@@ -2,7 +2,7 @@ import { transformAsNestedBaseTargetRecord } from '../../../../../utils/transfor
 import type { ColumnSchema } from '../../../../../sdlc/single-report-schema';
 import { SaferTableSchema, Selectable } from '../../../../../types';
 import { Box, Flex, Grid, GridItem, Icon, Text } from '@chakra-ui/react';
-import { getAssertions } from '../../../../../utils/assertion';
+import { getAssertions } from '../../utils';
 import { FiArrowRight, FiChevronRight } from 'react-icons/fi';
 import HistogramChart from '../../../Charts/HistogramChart';
 import { ColumnName } from '../ColumnName';
